Refactor app structure and cleanup
