+++ package.json
@@ -4,6 +4,7 @@
   "private": false,
   "dependencies": {
     "lucide-react": "^0.263.1",
+    "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1"
