+++ src/components/EndDayPhase.jsx
@@ -1,0 +1,42 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { Star } from 'lucide-react';
+
+const EndDayPhase = ({ gameState, startNewDay }) => (
+  <div className="bg-white rounded-lg shadow-lg p-4 max-w-2xl mx-auto">
+    <h2 className="text-xl font-bold mb-4 flex items-center gap-2">
+      <Star className="w-5 h-5" />
+      Day {gameState.day} Complete!
+    </h2>
+
+    <div className="grid grid-cols-2 gap-4 mb-6">
+      <div className="bg-green-100 p-3 rounded-lg text-center">
+        <h3 className="font-bold text-green-800 text-sm">Today's Earnings</h3>
+        <p className="text-lg font-bold text-green-600">
+          {gameState.customers.reduce((total, c) => total + (c.payment || 0), 0)} Gold
+        </p>
+      </div>
+
+      <div className="bg-blue-100 p-3 rounded-lg text-center">
+        <h3 className="font-bold text-blue-800 text-sm">Customers Served</h3>
+        <p className="text-lg font-bold text-blue-600">
+          {gameState.customers.filter(c => c.satisfied).length} / {gameState.customers.length}
+        </p>
+      </div>
+    </div>
+
+    <button
+      onClick={startNewDay}
+      className="w-full bg-amber-500 hover:bg-amber-600 text-white py-3 rounded-lg font-bold text-lg"
+    >
+      Start Day {gameState.day + 1}
+    </button>
+  </div>
+);
+
+EndDayPhase.propTypes = {
+  gameState: PropTypes.object.isRequired,
+  startNewDay: PropTypes.func.isRequired,
+};
+
+export default EndDayPhase;
